# Prevent multiple allocations of ReaderParameters