`ReaderParams` gets allocated multiple times inside `DirectoryAssemblyResolver`, accounting for about 4.6% of the memory allocated in the process when linking a hello world app. Preventing this many allocations can help us improve the performance of the linker just a little bit by easing up the GC work.

![image](https://user-images.githubusercontent.com/3836488/92987590-b9e0ed00-f478-11ea-8283-0f48581b224f.png)

Comparison table (linking a hello world app)

|                      | Before   | After    |
|----------------------|----------|----------|
| Total Allocs            | 324MB | 305MB |
| Total CPU Time       | 20,642ms | 19,672ms |
| Total GC CPU Time    | 4,546ms  | 4,235ms  |
| Total GC Pause       | 4,795ms  | 4,468ms  |
| % Time paused for GC | 24%      | 23.2%    |

/cc @agocke